Raised the API version to v36.0 to allow for access to the Organization record's field : NamespacePrefix
Added ApexClass.CreatedBy.Name and ApexClass.LastModifiedBy.Name to the list of getSObjectFieldList() method and removed the need for the static string of ADDITIONAL_FIELDS.  Also adjust all queries accordingly for this change.

All queries now focus only on the org's default namespace prefix as determined by the Organization object.