Bugfix
* Resolved issue where a "refresh all” command can hit
"System.LimitException: Apex CPU time limit exceeded" - Change to
CodeDocumenter now moves the request out to a batch mode for better
distribution of documenting load.
* Implemented Salesforce Streaming API to monitor changes to the
CodeInfo table.  As the classes are processed in the background batch
threads, the Streaming API will alert the page to display the class
names as needed.