Various minor adjustments
* Variable name changes to better represent purpose
* In the main execute() method, the call to resolveTheClassLinks() now
only occurs if the execution thread is not operating in a batch or
queueable context.  This will allow the call to occur only after all
class documentation tasks are completed in a batch scenario.
* In the primary execute() method, I increased the details provided to
the nestedClazz construct to avoid an NPE.
* Corrected spelling of the CodeDocumenterException class
* The parseApexClass(ApexClass, Boolean) method will now re-throw any
exception encountered wrapped in a CodeDocumenterException instead of
swallowing the original exception.
* removed a debug statement