Removed the need to have model_type as part of the JSON model
This was originally needed when the model was stored in the separate
attachment.  The thought was that if the attachment record were
separated from the parent record in some context, then application code
would still know how to deserialize it.  That is not required now that
the JSON model is stored on the CodeInfo__c record directly.